Fix openSettings context access
I've mentioned to extension developers already, but it's probably good to follow here as well.

Rather than using context to get the activity outside of the callback, We have to use the version provided to the callback. Otherwise when there is an activity state change the context is outdated and the settings dialog won't open, or in some cases cause the app to crash with an exception.